[IMP] runbot: kill older builds when a new one is created
When a new commit is found and a new build is created. When a user
pushes more commits in the same branch a few minutes later, it causes
more builds in testing phase, resulting in a CPU waste.

With this commit, previous builds in testing phase in non sticky
branches are killed when a new build is created.